STEP2: 예외처리 및 리팩터링 (#70)
* STEP1: 1단계 예약 기능 구현

- test : 메서드 이름 명확하게 수정

* STEP1: 1단계 예약 기능 구현

- fix : Reservation id 에 대해서 equals, hashCode Override 추가

* STEP1: 2단계 테마 관리 기능 추가

- fix: html 내용 수정 및 sql 수정

* STEP1: 2단계 테마 관리 기능 추가

- feat: 테마 관리 기능에 필요한 js, html 추가

* STEP1: 2단계 테마 관리 기능 추가

- feat: 테마 관리 기능 추가

* STEP1: 2단계 테마 관리 기능 추가

- feat: 예외 처리 추가

* STEP2: 2단계 테마 관리 기능 추가

- fix : validateReservation 수정 및 분리
- fix : 불필요한 setter 제거
- fix : RoomEscapeException 이외의 다른 예외 처리 관리 추가
- fix : 불필요한 주석 제거

* STEP2: 예외처리 및 리팩터링

- fix: time.html 에 nav bar 추가
- fix: handleHttpMessageNotReadable 추가
- fix: isDuplicateReservation 수정
- fix: 예외 처리 추가

* STEP2: 예외처리 및 리팩터링

- fix: 에러코드 개선
- fix: 테마 예외처리 추가

* STEP2: 예외처리 및 리팩터링

- fix: 에러 응답 코드 추가

* STEP2: 예외처리 및 리팩터링

- fix: 예외처리 누락 파일 추가

* STEP2: 예외처리 및 리팩터링

- test: 불필요한 import 제거

+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,10 +1,15 @@
 package roomescape.service;
 
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 
 import roomescape.controller.dto.ReservationRequest;
 import roomescape.controller.dto.ReservationResponse;
 import roomescape.domain.Reservation;
+import roomescape.exception.ErrorCode;
+import roomescape.exception.RoomEscapeException;
 import roomescape.repository.ReservationRepository;
 
 import org.springframework.stereotype.Service;
@@ -16,27 +21,59 @@ public class ReservationService {
 
 	private final ReservationTimeService reservationTimeService;
 
-	ReservationService(ReservationRepository reservationRepository, ReservationTimeService reservationTimeService) {
+	private final ThemeService themeService;
+
+	ReservationService(ReservationRepository reservationRepository, ReservationTimeService reservationTimeService,
+			ThemeService themeService) {
 		this.reservationRepository = reservationRepository;
 		this.reservationTimeService = reservationTimeService;
+		this.themeService = themeService;
 	}
 
 	public List<ReservationResponse> getReservations() {
 		return this.reservationRepository.findAll()
 			.stream()
-			.map((reservation) -> ReservationResponse.from(reservation, reservation.getTime()))
+			.map((reservation) -> ReservationResponse.from(reservation, reservation.getTime(), reservation.getTheme()))
 			.toList();
 	}
 
 	public ReservationResponse create(ReservationRequest request) {
 		var reservationTime = this.reservationTimeService.getReservationTimeById(request.timeId());
-		var reservation = Reservation.builder().name(request.name()).date(request.date()).time(reservationTime).build();
+
+		checkReservationAvailability(request, reservationTime.getStartAt());
+
+		var theme = this.themeService.getThemeById(request.themeId());
+		var reservation = Reservation.builder()
+			.name(request.name())
+			.date(request.date())
+			.time(reservationTime)
+			.theme(theme)
+			.build();
 		var savedReservation = this.reservationRepository.save(reservation);
-		return ReservationResponse.from(savedReservation, reservationTime);
+		return ReservationResponse.from(savedReservation, reservationTime, theme);
 	}
 
 	public void cancel(long id) {
-		this.reservationRepository.delete(id);
+		var isExist = this.reservationRepository.isExistId(id);
+		if (isExist) {
+			this.reservationRepository.delete(id);
+		}
+		else {
+			throw new RoomEscapeException(ErrorCode.NOT_FOUND_RESERVATION);
+		}
+	}
+
+	private void checkReservationAvailability(ReservationRequest request, String time) {
+		LocalDate reservationDate = LocalDate.parse(request.date());
+		LocalTime reservationTime = LocalTime.parse(time, DateTimeFormatter.ofPattern("HH:mm"));
+		if (reservationDate.isBefore(LocalDate.now())
+				|| (reservationDate.isEqual(LocalDate.now()) && reservationTime.isBefore(LocalTime.now()))) {
+			throw new RoomEscapeException(ErrorCode.PAST_RESERVATION);
+		}
+
+		if (this.reservationRepository.isDuplicateReservation(request.date(), request.themeId())) {
+			throw new RoomEscapeException(ErrorCode.DUPLICATE_RESERVATION);
+		}
 	}
 
 }
